// src/app/models/admin/tinhtiencapquyen/tttientheonam.model.ts
export class OutputTtTinhTienTheoNamModel {
  idtientheonam: string;
  sotienhangnam: number;
  namnoptien: number;
  idtinhtiencapquyen: string;
  siteid: string;
}

export class InputTtTinhTienTheoNamModel {
  idtientheonam: string;
  sotienhangnam: number;
  namnoptien: number;
  idtinhtiencapquyen: string;
  siteid: string;
}